refactor(ffi-bridge): lift share_block proptest into its own test bin
Pure refactor: split the 526-line tests/share_block.rs into three
files per issue #36, sharing fixtures via the standard Cargo
tests/<subdir>/mod.rs pattern.

- tests/share_block_helpers/mod.rs (117) — fixture_folder,
   copy_dir_recursive, fresh_writable_vault, mint_external_card,
   save_one_record_block + the test UUID / timestamp constants.
   NOT a test bin (tests/<subdir>/ is sub-module territory); each
   consuming bin pulls it in via `mod share_block_helpers;`.
- tests/share_block.rs (347) — 7 integration tests covering the
   happy path + 6 typed-error failure modes.
- tests/share_block_proptest.rs (96) — the N-recipient sequential-
   share property test in its own bin with an independent
   case-budget (issue #38 still tracks raising it via shared fixture).

Test counts unchanged: 599 passed + 9 ignored (the proptest's 16 cases
move from one bin to another but stay in the cargo test run; the bin
split makes them independently runnable via `cargo test --test share_block_proptest`).
